refactor(memory): standardize cognitive record memory on kernel layout and remove tierstore
Standardize all cognitive record memory stores (WorkingRecordMemory, SemanticRecordMemory,
ProceduralRecordMemory, EpisodicPartitionedMemory) on the standard 64-byte Kernel MemoryHeader
(0x534D4B4D). Delete legacy TierStore interface and CognitiveRecordLayoutAdapter. Elevate
visibleCount(), isPersistent(), tombstoneRatio(), force(), and cognitiveLayout() onto
CognitiveRecordMemory and RecordMemory<L>, completely eliminating instanceof downcasting
across TierRouter, ConsolidationService, DuplicateDetector, VacuumCompactor, GraphExpansionStage,
and DefaultSpectorMemory. Migrate MemoryIndex to IndexRecordMemory extending AbstractRecordMemory
and EpisodicMemoryStore to EpisodicPartitionedMemory implementing PartitionedRecordMemory<L>.

Closes #380, #381, #386
