+++ memory/spector-memory/src/main/java/com/spectrayan/spector/memory/sync/VacuumCompactor.java
@@ -12,7 +12,7 @@
  */
 package com.spectrayan.spector.memory.sync;
 
-import com.spectrayan.spector.memory.cortex.AbstractCognitiveRecordMemory;
+import com.spectrayan.spector.memory.cortex.CognitiveRecordMemory;
 import com.spectrayan.spector.memory.index.MemoryIndex;
 import com.spectrayan.spector.memory.model.MemoryType;
 import com.spectrayan.spector.memory.synapse.CognitiveRecordLayout;
@@ -75,13 +75,13 @@ public final class VacuumCompactor {
      * @param index   the memory index (for offset remapping)
      * @return the compaction result (null if no compaction needed)
      */
-    public static CompactionResult compact(AbstractCognitiveRecordMemory store, MemoryType type,
+    public static CompactionResult compact(CognitiveRecordMemory store, MemoryType type,
                                             MemoryIndex index) {
         long startMs = System.currentTimeMillis();
 
         CognitiveRecordLayout layout = store.cognitiveLayout();
         int totalRecords = store.size();
-        long baseOffset = store.isPersistent() ? AbstractCognitiveRecordMemory.METADATA_HEADER_BYTES : 0;
+        long baseOffset = store.isPersistent() ? CognitiveRecordMemory.METADATA_HEADER_BYTES : 0;
         int stride = layout.stride();
 
         // Phase 1: Count live and tombstoned records
@@ -165,7 +165,7 @@ public final class VacuumCompactor {
      * @param threshold the tombstone ratio threshold (e.g., 0.20 for 20%)
      * @return true if compaction is recommended
      */
-    public static boolean shouldCompact(AbstractCognitiveRecordMemory store, float threshold) {
+    public static boolean shouldCompact(CognitiveRecordMemory store, float threshold) {
         if (store.size() == 0) return false;
         return store.tombstoneRatio() >= threshold;
     }
