+++ memory/spector-memory/src/test/java/com/spectrayan/spector/memory/index/MemoryIndexTest.java
@@ -14,7 +14,7 @@ package com.spectrayan.spector.memory.index;
 
 import com.spectrayan.spector.memory.model.MemoryType;
 import com.spectrayan.spector.memory.cortex.MemorySource;
-import com.spectrayan.spector.memory.index.MemoryIndex.MemoryLocation;
+import com.spectrayan.spector.memory.index.IndexRecordMemory.MemoryLocation;
 
 import org.junit.jupiter.api.*;
 import static org.assertj.core.api.Assertions.*;
